refactor: deduplicate Lua items utilities
